requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
pymongo==4.6.1
flask-jwt-extended==4.6.0
bcrypt==4.1.2
python-jose==3.3.0
openai-whisper==20231117
google-generativeai==0.3.1
python-multipart==0.0.6
yt-dlp==2023.12.30
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
crewai==0.83.0
crewai-tools==0.14.0
